fix: only kept theme option for the preview component and remove it from everywhere else
