+++ package.json
@@ -17,7 +17,6 @@
     "@ai-sdk/react": "^2.0.18",
     "@codemirror/lang-css": "^6.3.1",
     "@codemirror/lang-javascript": "^6.2.4",
-    "@codemirror/theme-one-dark": "^6.1.3",
     "@next/third-parties": "^15.5.2",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
@@ -34,6 +33,7 @@
     "@tsparticles/react": "^3.0.0",
     "@tsparticles/slim": "^3.9.1",
     "@types/inquirer": "^9.0.9",
+    "@uiw/codemirror-theme-vscode": "^4.25.1",
     "@uiw/react-codemirror": "^4.25.1",
     "ai": "^5.0.18",
     "class-variance-authority": "^0.7.1",
