+++ src/components/playground/playground-context.tsx
@@ -2,7 +2,7 @@
 
 import { createContext, useContext, useState, useCallback, ReactNode } from "react";
 import { ComponentConfig, PlaygroundState } from "./types";
-import { Theme } from "@/lib/themes";
+
 
 function parseJSXProps(code: string): Record<string, any> {
   try {
@@ -68,8 +68,6 @@ interface PlaygroundContextType {
   exportCode: () => string;
   copyCode: () => Promise<void>;
   updateStyles: (styles: string) => void;
-  setTheme: (theme: Theme) => void;
-  setDarkMode: (isDark: boolean) => void;
 }
 
 const PlaygroundContext = createContext<PlaygroundContextType | undefined>(undefined);
@@ -79,8 +77,6 @@ export function PlaygroundProvider({ children }: { children: ReactNode }) {
     selectedComponent: null,
     code: "",
     props: {},
-    theme: "default",
-    isDarkMode: true,
     styles: `/* Component styles */
 .component-container {
   /* Add your custom styles here */
@@ -150,13 +146,7 @@ export function PlaygroundProvider({ children }: { children: ReactNode }) {
     setState((prev: PlaygroundState) => ({ ...prev, styles }));
   }, []);
 
-  const setTheme = useCallback((theme: Theme) => {
-    setState((prev: PlaygroundState) => ({ ...prev, theme: theme.name }));
-  }, []);
 
-  const setDarkMode = useCallback((isDark: boolean) => {
-    setState((prev: PlaygroundState) => ({ ...prev, isDarkMode: isDark }));
-  }, []);
 
   const value: PlaygroundContextType = {
     state,
@@ -167,8 +157,6 @@ export function PlaygroundProvider({ children }: { children: ReactNode }) {
     exportCode,
     copyCode,
     updateStyles,
-    setTheme,
-    setDarkMode,
   };
 
   return (
