+++ src/components/playground/preview-panel.tsx
@@ -2,22 +2,74 @@
 
 import React, { useEffect, useState } from "react";
 import { usePlayground } from "./playground-context";
-import { useTheme } from "@/contexts/theme-context";
+import { ThemeProvider, useTheme } from "@/contexts/theme-context";
 import { Button } from "@/components/ui/button";
-import { RefreshCw, Maximize2, Minimize2, Monitor, Smartphone } from "lucide-react";
+import { 
+  DropdownMenu, 
+  DropdownMenuContent, 
+  DropdownMenuItem, 
+  DropdownMenuTrigger,
+  DropdownMenuSeparator 
+} from "@/components/ui/dropdown-menu";
+import { RefreshCw, Maximize2, Minimize2, Monitor, Smartphone, Palette, Moon, Sun } from "lucide-react";
 import { cn } from "@/lib/utils";
+import { themes, applyScopedTheme } from "@/lib/themes";
 
 type ViewportSize = "desktop" | "mobile";
 
-export function PreviewPanel() {
+
+
+function ScopedThemeWrapper({ 
+  children, 
+  theme, 
+  isDark 
+}: { 
+  children: React.ReactNode; 
+  theme: typeof themes[0]; 
+  isDark: boolean; 
+}) {
+  const ref = React.useRef<HTMLDivElement>(null);
+  
+  React.useEffect(() => {
+    if (ref.current) {
+      const vars = isDark ? theme.cssVars.dark : theme.cssVars.light;
+      Object.keys(vars).forEach(key => {
+        ref.current?.style.removeProperty(key);
+      });
+      
+     
+      applyScopedTheme(ref.current, theme, isDark);
+      
+    }
+  }, [theme, isDark]);
+  
+  return (
+    <div 
+      ref={ref}
+      className={cn(isDark ? "dark" : "", "h-full w-full")}
+      style={{
+
+        backgroundColor: isDark ? theme.cssVars.dark['--background'] : theme.cssVars.light['--background'],
+        color: isDark ? theme.cssVars.dark['--foreground'] : theme.cssVars.light['--foreground'],
+        minHeight: '100%',
+        width: '100%'
+      }}
+    >
+      {children}
+    </div>
+  );
+}
+
+function PreviewPanelContent() {
   const { state } = usePlayground();
-  const { previewDarkMode, currentTheme } = useTheme();
+  const { previewDarkMode, currentTheme, setTheme, setPreviewDarkMode } = useTheme();
   const [error, setError] = useState<string | null>(null);
   const [Component, setComponent] = useState<React.ComponentType<any> | null>(null);
   const [viewportSize, setViewportSize] = useState<ViewportSize>("desktop");
   const [isFullscreen, setIsFullscreen] = useState(false);
   const [refreshKey, setRefreshKey] = useState(0);
   const [isRefreshing, setIsRefreshing] = useState(false);
+  const [isThemeMenuOpen, setIsThemeMenuOpen] = useState(false);
 
   useEffect(() => {
     if (state.selectedComponent) {
@@ -26,25 +78,9 @@ export function PreviewPanel() {
     }
   }, [state.selectedComponent]);
 
+  
   useEffect(() => {
-    const root = document.documentElement;
-    if (previewDarkMode) {
-      root.classList.add("dark");
-    } else {
-      root.classList.remove("dark");
-    }
-  }, [previewDarkMode]);
-
-  // Apply theme CSS variables and custom styles
-  useEffect(() => {
-    const root = document.documentElement;
-    const cssVars = previewDarkMode ? currentTheme.cssVars.dark : currentTheme.cssVars.light;
     
-    Object.entries(cssVars).forEach(([property, value]) => {
-      root.style.setProperty(property, value);
-    });
-
-    // Apply custom styles
     let styleElement = document.getElementById('playground-custom-styles');
     if (!styleElement) {
       styleElement = document.createElement('style');
@@ -52,13 +88,13 @@ export function PreviewPanel() {
       document.head.appendChild(styleElement);
     }
     styleElement.textContent = state.styles;
-  }, [currentTheme, previewDarkMode, state.styles]);
+  }, [state.styles]);
 
   const handleRefresh = async () => {
     setIsRefreshing(true);
     setError(null);
     
-    // Small delay to show the refresh animation
+    
     await new Promise(resolve => setTimeout(resolve, 100));
     
     setRefreshKey(prev => prev + 1);
@@ -76,9 +112,18 @@ export function PreviewPanel() {
     }
   };
 
+  const handleThemeSelect = (theme: typeof themes[0]) => {
+    setTheme(theme);
+    setIsThemeMenuOpen(false);
+  };
+
+  const toggleDarkMode = () => {
+    setPreviewDarkMode(!previewDarkMode);
+  };
+
   if (!state.selectedComponent) {
     return (
-      <div className="h-full flex items-center justify-center bg-background text-muted-foreground">
+      <div className="h-full flex items-center justify-center bg-background text-muted-foreground dark">
         <div className="text-center">
           <h3 className="text-lg font-semibold mb-2 text-foreground">PREVIEW</h3>
           <p className="text-sm">Select a component to see it in action</p>
@@ -89,10 +134,10 @@ export function PreviewPanel() {
 
   return (
     <div className={cn(
-      "h-full flex flex-col bg-background",
+      "h-full flex flex-col bg-background dark",
       isFullscreen && "fixed inset-0 z-50"
     )}>
-      {/* Preview Header */}
+      
       <div className="flex items-center justify-between px-4 py-2 border-b border-border bg-muted/50">
         <div className="flex items-center gap-2">
           <span className="text-sm text-muted-foreground">Live Preview</span>
@@ -107,7 +152,7 @@ export function PreviewPanel() {
         </div>
         
         <div className="flex items-center gap-1">
-          {/* Viewport Controls */}
+          
           <div className="flex items-center gap-1 mr-2">
             <Button
               size="sm"
@@ -131,6 +176,50 @@ export function PreviewPanel() {
             >
               <Monitor className="h-4 w-4" />
             </Button>
+          </div>
+
+          
+          <div className="flex items-center gap-1 mr-2">
+            <Button
+              size="sm"
+              variant="outline"
+              onClick={toggleDarkMode}
+              className="h-8 w-8 p-0"
+            >
+              {previewDarkMode ? (
+                <Sun className="h-4 w-4" />
+              ) : (
+                <Moon className="h-4 w-4" />
+              )}
+            </Button>
+            
+            <DropdownMenu open={isThemeMenuOpen} onOpenChange={setIsThemeMenuOpen}>
+              <DropdownMenuTrigger asChild>
+                <Button size="sm" variant="outline" className="h-8">
+                  <Palette className="h-4 w-4 mr-2" />
+                </Button>
+              </DropdownMenuTrigger>
+              <DropdownMenuContent align="end" className="w-48">
+                <div className="px-2 py-1.5 text-sm font-semibold text-muted-foreground">
+                  Themes
+                </div>
+                <DropdownMenuSeparator />
+                {themes.map((theme) => (
+                  <DropdownMenuItem
+                    key={theme.name}
+                    onClick={() => handleThemeSelect(theme)}
+                    className="flex items-center justify-between"
+                  >
+                    <div className="flex items-center gap-2">
+                      <span>{theme.label}</span>
+                    </div>
+                    {currentTheme.name === theme.name && (
+                      <div className="w-2 h-2 rounded-full bg-primary" />
+                    )}
+                  </DropdownMenuItem>
+                ))}
+              </DropdownMenuContent>
+            </DropdownMenu>
           </div>
 
           <Button
@@ -158,49 +247,51 @@ export function PreviewPanel() {
         </div>
       </div>
 
-      {/* Preview Content */}
-      <div className="flex-1 overflow-auto bg-background">
-        <div className="min-h-full flex items-center justify-center p-4">
-          {error ? (
-            <div className="text-center p-8">
-              <div className="text-destructive text-lg font-semibold mb-2">
-                Component Error
+      
+      <ScopedThemeWrapper theme={currentTheme} isDark={previewDarkMode}>
+        <div className="flex-1 overflow-auto">
+          <div className="min-h-full flex items-center justify-center p-4">
+            {error ? (
+              <div className="text-center p-8">
+                <div className="text-destructive text-lg font-semibold mb-2">
+                  Component Error
+                </div>
+                <div className="text-muted-foreground mb-4">{error}</div>
+                <Button onClick={handleRefresh} variant="outline" size="sm">
+                  Try Again
+                </Button>
               </div>
-              <div className="text-muted-foreground mb-4">{error}</div>
-              <Button onClick={handleRefresh} variant="outline" size="sm">
-                Try Again
-              </Button>
-            </div>
-          ) : Component ? (
-            <div 
-              className={cn(
-                "w-full transition-all duration-300 ease-in-out",
-                viewportSize === "mobile" ? "max-w-sm" : "max-w-6xl"
-              )}
-              style={{ 
-                maxWidth: getViewportWidth(),
-                minHeight: viewportSize === "mobile" ? "600px" : "auto"
-              }}
-            >
-              <div className="w-full h-full flex items-center justify-center">
-                <ErrorBoundary key={refreshKey} onError={setError}>
-                  <Component key={refreshKey} {...state.props} />
-                </ErrorBoundary>
+            ) : Component ? (
+              <div 
+                className={cn(
+                  "w-full transition-all duration-300 ease-in-out",
+                  viewportSize === "mobile" ? "max-w-sm" : "max-w-6xl"
+                )}
+                style={{ 
+                  maxWidth: getViewportWidth(),
+                  minHeight: viewportSize === "mobile" ? "600px" : "auto"
+                }}
+              >
+                <div className="w-full h-full flex items-center justify-center">
+                  <ErrorBoundary key={refreshKey} onError={setError}>
+                    <Component key={refreshKey} {...state.props} />
+                  </ErrorBoundary>
+                </div>
               </div>
-            </div>
-          ) : (
-            <div className="text-center text-muted-foreground">
-              <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-primary mx-auto mb-2"></div>
-              <p>Loading component...</p>
-            </div>
-          )}
+            ) : (
+              <div className="text-center text-muted-foreground">
+                <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-primary mx-auto mb-2"></div>
+                <p>Loading component...</p>
+              </div>
+            )}
+          </div>
         </div>
-      </div>
+      </ScopedThemeWrapper>
     </div>
   );
 }
 
-// Error Boundary for Component Preview
+
 class ErrorBoundary extends React.Component<
   { children: React.ReactNode; onError: (error: string) => void },
   { hasError: boolean }
@@ -234,4 +325,12 @@ class ErrorBoundary extends React.Component<
 
     return this.props.children;
   }
+}
+
+export function PreviewPanel() {
+  return (
+    <ThemeProvider>
+      <PreviewPanelContent />
+    </ThemeProvider>
+  );
 }
